+++ src/app/ubs-admin/ubs-admin-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { UbsUserOrdersComponent } from './components/ubs-user-orders/ubs-user-orders.component';
+import { UbsUserOrdersComponent } from '../ubs-user/ubs-user-orders/ubs-user-orders.component';
 import { UbsAdminEmployeeComponent } from './components/ubs-admin-employee/ubs-admin-employee.component';
 import { UbsClientProfilePageComponent } from './components/ubs-client-profile-page/ubs-client-profile-page.component';
 import { UbsAdminTableComponent } from './components/ubs-admin-table/ubs-admin-table.component';
@@ -15,7 +15,6 @@ const ubsAdminRoutes: Routes = [
     component: UbsAdminComponent,
     canActivate: [UbsAdminGuardGuard],
     children: [
-      { path: '', component: UbsUserOrdersComponent },
       { path: 'orders', component: UbsAdminTableComponent },
       { path: 'employee/:page', component: UbsAdminEmployeeComponent },
       { path: 'profile', component: UbsClientProfilePageComponent },
