+++ src/app/ubs-user/ubs-user-order-details/ubs-user-order-details.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { UserOrdersService } from '../../services/user-orders.service';
+import { UserOrdersService } from '../../ubs-user/services/user-orders.service';
 import { takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
