+++ src/app/ubs-user/ubs-user-routing.module.ts
@@ -2,12 +2,14 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { UbsUserGuardGuard } from './ubs-user-guard.guard';
 import { UbsUserComponent } from './ubs-user.component';
-
+import { UbsUserOrdersComponent } from './ubs-user-orders/ubs-user-orders.component';
 const routes: Routes = [
   {
     path: '',
     component: UbsUserComponent,
-    canActivate: [UbsUserGuardGuard]
+    canActivate: [UbsUserGuardGuard],
+
+    children: [{ path: 'orders', component: UbsUserOrdersComponent }]
   }
 ];
